первый драфт

+++ internal/repository/migration/0002_create_order_table.up.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS orders
     user_id uuid NOT NULL,
     "number" character varying NOT NULL,
     status character varying NOT NULL,
-    accrual money,
+    accrual float,
     created_at timestamp without time zone NOT NULL,
     updated_at timestamp without time zone NOT NULL
 )
