+++ internal/repository/migration/0004_create_user_withdrawals_balance_table.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS user_withdrawals
 (
     id uuid NOT NULL,
     user_id uuid NOT NULL,
-    sum money,
+    sum float,
     "number" character varying NOT NULL,
     created_at timestamp
 )
